requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.2
django-allauth==0.41.0
django-ckeditor==5.9.0
django-countries==6.0
django-crispy-forms==1.8.1
django-debug-toolbar==2.2
django-filter==2.2.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-progressive-web-app==0.1.1
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
docopt==0.6.2
gitignore==0.0.8
gunicorn==20.0.4
heroku==0.1.4
idna==2.8
Markdown==3.2.1
oauthlib==3.1.0
Pillow==7.0.0
psycopg2==2.8.4
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.9.3
python-dateutil==1.5
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
soupsieve==1.9.5
sqlparse==0.3.0
static3==0.7.0
stripe==2.43.0
urllib3==1.25.8
whitenoise==5.0.1
xlwt==1.3.0
